pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aat"
version = "0.1.0"
description = "Local Academic Translation Agent (EN→ZH), Reference-Aware, Free Search by Default"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "python-docx>=1.1.0",
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "anthropic>=0.40.0",
    "openai>=1.10.0",
    "httpx>=0.26.0",
    "chromadb>=0.4.0",
    "ollama>=0.2.0",
    "tiktoken>=0.5.0",
    "toml>=0.10.2",
    "rich>=13.0.0",
    "tqdm>=4.66.0",
    "fastapi>=0.110.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "hypothesis>=6.92.0",
    "ruff>=0.1.0",
    "black>=23.12.0",
    "mypy>=1.8.0",
]

[project.scripts]
aat = "aat.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
    "--strict-markers",
    "--disable-warnings",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "property: Property tests using hypothesis",
]

[tool.coverage.run]
source = ["aat"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 85
show_missing = true

[tool.coverage.file]
# Critical files that require 95% coverage
fail_under = [
    "aat/translate/validators.py = 95",
    "aat/parsing/citation.py = 95",
    "aat/translate/pipeline.py = 95",
    "aat/orchestrator/graph.py = 95",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = [
    "tests/",
]
